Switch to Github script
