--- .github/actions/changelog/.eslintrc.js
@@ -1,17 +1,0 @@
-module.exports = {
-  env: {
-    browser: true,
-    es2021: true
-  },
-  extends: 'standard',
-  plugins: ["prettier"],
-  overrides: [
-  ],
-  parserOptions: {
-    ecmaVersion: 'latest',
-    sourceType: 'module'
-  },
-  rules: {
-    "prettier/prettier": "error"
-  }
-}
