--- .github/actions/changelog/action.yml
@@ -1,5 +1,0 @@
-name: 'Changelog'
-description: 'Check if the CHANGELOG entry added.'
-runs:
-  using: 'node16'
-  main: 'dist/index.js'
