+++ .github/actions/changelog/package.json
@@ -2,10 +2,9 @@
   "name": "ChangelogChecker",
   "version": "1.0.0",
   "description": "Makes sure that the CHANGELOG.md has the entry for the current pull request.",
-  "main": "changelog.js",
+  "main": "index.js",
   "scripts": {
-    "format": "prettier-eslint --write changelog.js",
-    "release": "ncc build changelog.js"
+    "format": "prettier --write index.js"
   },
   "repository": {
     "type": "git",
@@ -22,16 +21,6 @@
     "@actions/github": "^5.1.1"
   },
   "devDependencies": {
-    "@vercel/ncc": "^0.34.0",
-    "eslint": "^8.29.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-n": "^15.6.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.1.1",
-    "prettier": "2.8.0",
-    "prettier-eslint": "^15.0.1",
-    "prettier-eslint-cli": "^7.1.0"
+    "prettier": "2.8.0"
   }
 }
